Do not input in db if the version already exists
On machines that took an upgrade with this number before, this upgrade
would return a non-zero value, and not write the flag file.
Thus, before inserting into db, we check that the version number
exists.
At this point, if the number exists, it means the machine was already
updated. If not, it just was from the script