Change the language column size
During the migration to the new language system, it was noticed that the
language field in the database, in `user_pref` and `domain_pref` was not
long enough for the ISO language codes.
This update increases the languages field to 10 characters.
Also adding a library to check that the fields are actually added to the
database properly.