Upgrade to Exim 4.94.2
Script should provide smooth upgrade as soon as the new package becomes available.

Safely prevents upgrade if main Git tree is out-of-date (missing new templates).

Safely prevents upgrade if the old templates are dirty since these changes will not be replicated.

Allows for installation to be forced when script is manually run with '--force' after the previous warning is accounted for.

Will run repeatedly until new package is successfully installed.