Update version.def daily instead of once
Historically, $SRCDIR/etc/mailcleaner/version.def was tracked by Git. This often results in it being marked as changed, even though it should not be tracked anymore. As a result, it can get reverted to the initial value (2021.02) when git is reset.

Instead, this commit dumps the version to the file after updates complete every evening.

Because of this, the value is just a duplicate of the longer value from the database, so this is redundant. For the future, we should consider showing a commit version. This iwll be largely meaningless to users, but would provide a quick way of checking if the Git tree is actually up to date. Currently we are just reporting that the update succeeded at one point in time. If a machine has been rolled back, it would then have incorrect information. This would not be true if we showed the most recent commit hash.