Prevent unnecessary restarts and change version to git commit
Only restart services if one of the Git trees changed, or if an update was successfully installed.

Since the commit hash is stored a variable already, we might as well start to use it as a non-redundant piece of version information.

TODO: In MC, add a SNMP target to get update status by comparing this value to the latest upstream commit. Display "Update available" when they differ.

Eventually: we'll have Community edition running 'master' and Enterprise running 'stable'. We'll need to build in branch-awareness.