Rebase Jessie machines to Updater4MC8
This is not tested and not ready to merge. Just an initial commit to
remove all other updates and create a basic outline. The rebase should
fail since Updater4MC8 has diverged from this. Will probably need to
reset to the last common commit for rebase to work.

TODO: Test